Sunteck Beach Residences Vasai West Mumbai
Welcome to Sunteck Suruchi Beach, an exclusive development in a residential segment that brings to reality the most ideal and looked up to the standard of living, and holds the actual meaning of a 'peaceful lifestyle'. Situated in one of the most beneficial and prominent locations of this city, at Vasai West Mumbai, Sunteck Suruchi Beach Vasai West is a gorgeous compilation of exclusive 2 & 3 BHK Ultra Luxurious Bed Residences that offer space and solitude to you and your people, right in the heart of the city. This venture is like dreams come true in 50 acres of land with 80% open area that will add the fragrance of luxury living to your every single day.

Sunteck Suruchi Beach Vasai Mumbai with the delightful exceptional infrastructure making available an actually unmatched way of life and ground-breaking advantages to the residents appears as a supreme chance to make every face happy with complete satisfaction.

Strategically located in Vasai West, Mumbai, investment in this residential development is the smart decision for a life-loving person like you who seeks to take his standard of living to newer heights. It presents the best of both worlds - metropolitan amenities and the friendship of the environment. Being the new booming destination of Mumbai, this location nearby Mira Road and Beyond, Mumbai, Palgarh District, Maharashtra. Nalasopara West (7.58 Km)& future metro station offers all the city conveniences and also good connectivity to Pune. Nestled in the setting of a green surrounding Sunteck Realty Project At Vasai West enjoys a green cover too.

Yes, Sunteck Realty through its residential project Suruchi Beach Vasai West appears as a well-wisher for the investors & home seekers by offering a chance for first-time home buyers with a long chain of advantages. So, what are you looking for? Every possible facility & feature available here get the premium values of living higher than other residential development in Mumbai in concern of up-to-date standard of living additionally.

For More Details :
Visit - https://sunteck.developerprojects.in/mumbai/sunteck-beach-residences-vasai-west/